rate_limiter/Backend/requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
python-dotenv